requirements.txt
requests>=2.0
python-dotenv>=1.0.0
langchain>=0.1.0
langgraph>=0.0.30